refactor: relocate agent-init.ts → src/agent/init.ts
Pure file move + relative import path updates inside the moved file
(12 paths) + 1 importer (legacy/core/runtime.ts) updated.

Part of #632 segment 3.
